AI-driven-Document-analysis-pipeline-main/AI-driven-Document-analysis-pipeline/requirements.txt
torch
transformers
pytesseract
PyPDF2
mysql-connector-python
torchcrf
pandas==1.5.2
sqlalchemy==2.0.0
torch==2.0.0
numpy==1.24.0
torchcrf==0.3.2
scikit-learn==1.0.1
pdfplumber
pytesseract 
pillow 
pandas
torch==1.13.0
transformers==4.11.3
pytorch-crf==0.7
pandas==1.3.3
PyPDF2==1.26.0
pytesseract==0.3.8
Pillow==8.4.0
scikit-learn==1.0.1

